:name_badge: Rename Receivers
