+++ Assets/Scripts/Receiver/EmptyReceiver.cs
@@ -1,6 +1,6 @@
-namespace UniFlow
+namespace UniFlow.Receiver
 {
-    public class EmptyReceiver : Receiver
+    public class EmptyReceiver : ReceiverBase
     {
         public override void OnReceive(EventMessages eventMessages)
         {
