+++ Assets/Scripts/ReceiverBase.cs
@@ -1,6 +1,6 @@
 namespace UniFlow
 {
-    public abstract class Receiver : EventConnectable, IReceiver
+    public abstract class ReceiverBase : EventConnectable, IReceiver
     {
         public abstract void OnReceive(EventMessages eventMessages);
     }
